requirements.txt
streamlit==1.42.0
PyPDF2==3.0.1
transformers==4.49.0
torch==2.6.0
google-generativeai==0.3.2
python-dotenv==1.0.1
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
